+++ intask/urls.py
@@ -6,5 +6,6 @@ urlpatterns = [
     url(r'^api/v1/users/', include('users.urls')),
     url(r'^api/v1/docs/', include('rest_framework_swagger.urls')),
     url(r'^api/v1/', include('projects.urls')),
+    url(r'^api/v1/', include('tasks.urls')),
     url(r'^', TemplateView.as_view(template_name='index.html')),
 ]
